Embed version information into executable
Allows use of AmigaDOS 'version' command to see version information
about PT-1210.

This will help cross-reference which git revision an executable was
built from.

Due to the Year 2000 bug, the 'version' command may not show the date
properly or at all unless you are using a fixed version of 'version' or
AmigaOS versions > 3.1.