CIA: Switch to continuous mode (fixes tempo drift)
Instead of re-arming the timer from within the CIA interrupt handler, we
use the continuous mode of the CIA and only write to the registers if we
need to adjust the tempo.

This avoids introducing delay into the replayer and hopefully fixes the worst
of the tempo drift/inaccuracy problems.