sample pointer clear fix
after much hex viewing, we've dertermined that samples of 2 bytes are never actually included in a module.  as such we've been over reaching the sample pointers and clearing data past the buffer.  re-worked mt_init so this case is properly taken into account and tidied up the crap around it.