Fix various module title printing bugs
Previously, it was not possible for a song title to occupy all 30
characters of the title area because the second-last and last empty
spaces (due to misalignment) was used as the blank "LCD background"
template to bitwise-OR with the new characters being printed.

Additionally, if a song title was long enough to occupy the second-last
space, the LCD background template would be trashed, so when a new
module was loaded, the LCD background would look wrong.

Finally, the LCD background was misaligned with the characters by one
pixel.

This commit fixes all the above by:

  * Adding the empty LCD space as a special character in the large font.
  * Updating the UI_Type functions to use this special character instead
  of relying on an empty space in the HUD.
  * Fixing the HUD artwork by moving the LCD area one pixel to the right
  to align it wit hthe characters properly.
  * Updating the C code to allow a title to use all 30 spaces.